Added tests and docstring checks
