+++ requirements.txt
@@ -1,3 +1,4 @@
+pygame
 pyfrc
 robotpy-ctre
 robotpy-navx
