+++ robot.py
@@ -16,14 +16,7 @@ if __name__ == '__main__':
         #todo verify this causes no issues on robot
         wpilib.DriverStation._reset()
 
-        #patch simulation
-        #we update the simluation files to ours. If we update WPIlib these may break
-        import sim
-        import pyfrc.sim
-        import pyfrc.sim.pygame_joysticks
-        pyfrc.sim.SimUI = sim.ui.SimUI
-        pyfrc.sim.pygame_joysticks.UsbJoysticks = sim.pygame_joysticks.UsbJoysticks
     except Exception as err:
-            print("Failed to patch runtime. Error", err)
+            print("Failed to do extra setup. Error", err)
     
     wpilib.run(Robot,physics_enabled=True)
