+++ team3200/tests/pyfrc_test.py
@@ -1,0 +1,24 @@
+'''
+    This test module imports tests that come with pyfrc, and can be used
+    to test basic functionality of just about any robot.
+'''
+
+
+import os
+
+if os.getenv("DOC_STRING_CHECK"):
+    from pyfrc.tests.docstring_test import *
+else:
+    from pyfrc.tests import *
+    def two_plus(arg):
+        """
+        Simple sample function
+        """
+        return 2 + arg
+
+    def test_addition():
+        """
+        Simple test example
+        """
+        assert two_plus(2) == 4
+
